requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.4.0
aiosqlite==0.19.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.4
attrs==25.4.0
certifi==2025.10.5
fastapi==0.121.2
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
pillow==10.2.0
propcache==0.4.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
yarl==1.22.0
